backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.4
redis==5.0.1

# AI and Agents
crewai==0.1.0
openai==1.3.7
pinecone-client==2.2.4
langchain==0.0.350
langchain-openai==0.0.2

# Web Search
serper-dev==0.1.0

# Vector Operations
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
ebooklib==0.18.1

# Image Processing
Pillow==10.1.0
requests==2.31.0

# Utilities
tenacity==8.2.3
structlog==23.2.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8

# Email
aiosmtplib==3.0.1

# Monitoring
sentry-sdk[fastapi]==1.38.0 